requirements.txt
torch
torchvision
matplotlib
numpy
pandas
pillow
scikit-learn